Factor out request building/sending logic and errors into separate files
